+++ boundaries/views.py
@@ -1,3 +1,5 @@
+from __future__ import unicode_literals
+
 from django.contrib.gis.db import models
 from django.http import Http404
 from django.utils.translation import ugettext as _