+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * I18n support.
 * Add tests.
 
+Identified quirks:
+
+* The `shape`, `simple_shape` and `centroid` endpoints ignore the `pretty` parameter.
+* `format=apibrowser` ignores JSONP and the `pretty` parameter.
+
 ## 0.4.0 (2014-08-01)
 
 * Add `start_date` and `end_date` to boundary sets. (@jamesturk)